Win port

+++ src/helper/time_helper.cpp
@@ -2,12 +2,12 @@
 
 long time_epoch_seconds() noexcept
 {
-	return std::chrono::duration_cast<std::chrono::seconds>(
-		        std::chrono::system_clock::now().time_since_epoch()).count();
+	return static_cast<long>(std::chrono::duration_cast<std::chrono::seconds>(
+		        std::chrono::system_clock::now().time_since_epoch()).count());
 }
 
 long time_epoch_miliseconds() noexcept
 {
-	return std::chrono::duration_cast<std::chrono::milliseconds>(
-			        std::chrono::system_clock::now().time_since_epoch()).count();
+	return static_cast<long>(std::chrono::duration_cast<std::chrono::milliseconds>(
+			        std::chrono::system_clock::now().time_since_epoch()).count());
 }
